separated explore commands
